pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dataframe-sampler"
version = "0.3.0"
description = "Generate synthetic CSV-like tabular samples from pandas DataFrames."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
    { name = "Fabrizio Costa" },
]
dependencies = [
    "click>=8.0",
    "numpy>=1.21",
    "pandas>=1.3",
    "PyYAML>=6.0",
    "scikit-learn>=1.0",
    "scipy>=1.7",
    "seaborn>=0.12",
]

[project.optional-dependencies]
test = [
    "pytest>=7",
]
parquet = [
    "pyarrow>=14",
]
pynndescent = [
    "pynndescent>=0.5",
]
hnswlib = [
    "hnswlib>=0.8",
]
annoy = [
    "annoy>=1.17",
]
ann = [
    "pynndescent>=0.5",
    "hnswlib>=0.8",
    "annoy>=1.17",
]

[project.scripts]
dataframe-sampler = "dataframe_sampler.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
